refactor: changed natives to std
