+++ std/collections.c
