+++ std/string.c
